Refactor unit tests and add most of the codes
