+++ tests/codes/color_3d/test_color_3d_code.py
@@ -1,0 +1,10 @@
+import pytest
+from panqec.codes import Color3DCode
+from tests.codes.stabilizer_code_test import StabilizerCodeTest
+
+
+class TestColor3DCode(StabilizerCodeTest):
+
+    @pytest.fixture
+    def code(self):
+        return Color3DCode(6, 6, 6)
