+++ tests/decoders/test_deformed_rotated_sweepmatch.py
